implementa tratamento de exceções

+++ src/main/java/com/grupo7/DroneFeeder/Service/DroneService.java
@@ -4,6 +4,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.grupo7.DroneFeeder.Entity.Drone;
+import com.grupo7.DroneFeeder.Erro.DroneExistenteException;
+import com.grupo7.DroneFeeder.Erro.DroneNaoEncontrado;
 import com.grupo7.DroneFeeder.Repository.DroneRepository;
 
 @Service
@@ -13,6 +15,9 @@ public class DroneService {
   private DroneRepository repository;
   
   public Drone createOneDrone(Drone drone) {
+    if (drone.getId() != 0) {
+      throw new DroneExistenteException();
+    }
     Drone result = repository.save(drone); 
     return result;
   }
@@ -23,18 +28,31 @@ public class DroneService {
   }
   
   public Drone findDrone(long id) {
-    Drone result = repository.findById(id).get();
-    return result;
+    if (this.exiteDrone(id)) {
+      Drone result = repository.findById(id).get();
+      return result;      
+    }
+    throw new DroneNaoEncontrado();
   }
   
   public Drone updateOneDrone(long id, Drone drone) {
-    Drone result = repository.findById(id).get();
-    result.setEntrega(drone.getEntrega());
-    return repository.save(result);
+    if (this.exiteDrone(id)) {
+      Drone result = repository.findById(id).get();
+      result.setEntrega(drone.getEntrega());
+      return repository.save(result);      
+    }
+    throw new DroneNaoEncontrado();
   }
   
   public String deleteOneDrone(long id) {
-    repository.deleteById(id);
-    return "drone deletado com sucesso";
+    if (this.exiteDrone(id)) {
+      repository.deleteById(id);
+      return "drone deletado com sucesso";      
+    }
+    throw new DroneNaoEncontrado();
+  }
+  
+  public boolean exiteDrone(long id) {
+    return repository.existsById(id);
   }
 }
